+++ frontend/src/components/index.tsx
@@ -1,4 +1,3 @@
 export * from './layouts';
 export * from './common';
 export * from './error';
-export { default as AnswerListRecheckModal } from './AnswerListRecheckModal';
