+++ frontend/src/pages/ReviewWritingCardFromPage/components/AnswerListRecheckModal/components/ReviewCard/styles.ts
